Router mechanic
Treat template path and page path as separate entities in order to keep main html structure after refresh.

Template retention after refresh with history.state and fallback in case of absence of history.

Home is now a template and changed menu labels with addition of aria-label and dataset.

Reusable function for fetch method.